Split server device actions by capability
