+++ src/main/java/pl/grzeslowski/openhab/supla/internal/SuplaHandlerFactory.java
@@ -1,11 +1,10 @@
 package pl.grzeslowski.openhab.supla.internal;
 
 import static java.util.Collections.synchronizedMap;
+import static java.util.stream.Collectors.toSet;
 import static pl.grzeslowski.openhab.supla.internal.SuplaBindingConstants.*;
 
-import java.util.HashMap;
-import java.util.Hashtable;
-import java.util.Map;
+import java.util.*;
 import org.eclipse.jdt.annotation.NonNull;
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
@@ -13,8 +12,10 @@ import org.openhab.core.config.discovery.DiscoveryService;
 import org.openhab.core.thing.Bridge;
 import org.openhab.core.thing.Thing;
 import org.openhab.core.thing.ThingTypeUID;
+import org.openhab.core.thing.ThingUID;
 import org.openhab.core.thing.binding.BaseThingHandlerFactory;
 import org.openhab.core.thing.binding.BridgeHandler;
+import org.openhab.core.thing.binding.ThingActions;
 import org.openhab.core.thing.binding.ThingHandler;
 import org.openhab.core.thing.binding.ThingHandlerFactory;
 import org.osgi.framework.ServiceReference;
@@ -31,9 +32,10 @@ import pl.grzeslowski.openhab.supla.internal.server.handler.*;
 /** The {@link SuplaHandlerFactory} is responsible for creating things and thing handlers. */
 @Component(service = ThingHandlerFactory.class, immediate = true, configurationPid = "binding.supla")
 @NonNullByDefault
-public class SuplaHandlerFactory extends BaseThingHandlerFactory {
+public class SuplaHandlerFactory extends BaseThingHandlerFactory implements ServerDeviceActionServiceRegistry {
     private final Logger logger = LoggerFactory.getLogger(SuplaHandlerFactory.class);
     private final Map<BridgeHandler, ServiceReference<?>> servicesToDispose = synchronizedMap(new HashMap<>());
+    private final Map<ThingUID, ActionServiceRegistrations> actionServicesToDispose = synchronizedMap(new HashMap<>());
 
     @Override
     public boolean supportsThingType(ThingTypeUID thingTypeUID) {
@@ -73,6 +75,9 @@ public class SuplaHandlerFactory extends BaseThingHandlerFactory {
 
     @Override
     protected void removeHandler(ThingHandler thingHandler) {
+        if (thingHandler instanceof ServerSuplaDeviceHandler serverHandler) {
+            unregisterActionServices(serverHandler);
+        }
         if (!(thingHandler instanceof BridgeHandler bridgeHandler)) {
             return;
         }
@@ -85,7 +90,7 @@ public class SuplaHandlerFactory extends BaseThingHandlerFactory {
 
     @NonNull
     private ThingHandler newServerDeviceHandler(final Thing thing) {
-        return new SingleDeviceHandler(thing);
+        return new SingleDeviceHandler(thing, this);
     }
 
     @NonNull
@@ -129,7 +134,7 @@ public class SuplaHandlerFactory extends BaseThingHandlerFactory {
 
     private ThingHandler newGatewayDeviceHandler(Thing thing) {
         var discovery = new ServerDiscoveryService(thing.getUID());
-        var bridgeHandler = new GatewayDeviceHandler(thing, discovery);
+        var bridgeHandler = new GatewayDeviceHandler(thing, discovery, this);
         var serviceRegistration = registerThingDiscovery(discovery);
         servicesToDispose.put(bridgeHandler, serviceRegistration.getReference());
         return bridgeHandler;
@@ -137,5 +142,61 @@ public class SuplaHandlerFactory extends BaseThingHandlerFactory {
 
     private ThingHandler newSubDeviceHandler(Thing thing) {
         return new SubDeviceHandler(thing);
+    }
+
+    @Override
+    public void updateActionServices(
+            ServerSuplaDeviceHandler handler, Set<Class<? extends ThingActions>> actionServices) {
+        var thingUID = handler.getThing().getUID();
+        var existing = actionServicesToDispose.get(thingUID);
+        if (existing != null && existing.actionServices().equals(actionServices)) {
+            return;
+        }
+
+        unregisterActionServices(handler);
+        var registrations = actionServices.stream()
+                .map(actionService -> registerActionService(handler, actionService))
+                .toList();
+        actionServicesToDispose.put(thingUID, new ActionServiceRegistrations(registrations));
+    }
+
+    @Override
+    public void unregisterActionServices(ServerSuplaDeviceHandler handler) {
+        var registrations = actionServicesToDispose.remove(handler.getThing().getUID());
+        if (registrations == null) {
+            return;
+        }
+        registrations.registrations().forEach(ActionServiceRegistration::unregister);
+    }
+
+    private ActionServiceRegistration registerActionService(
+            ServerSuplaDeviceHandler handler, Class<? extends ThingActions> actionService) {
+        try {
+            var service = actionService.getConstructor().newInstance();
+            service.setThingHandler(handler);
+            var registration = bundleContext.registerService(ThingActions.class, service, new Hashtable<>());
+            service.initialize();
+            return new ActionServiceRegistration(actionService, service, registration);
+        } catch (ReflectiveOperationException ex) {
+            throw new IllegalStateException("Could not register action service " + actionService.getName(), ex);
+        }
+    }
+
+    private record ActionServiceRegistrations(List<ActionServiceRegistration> registrations) {
+        private Set<Class<? extends ThingActions>> actionServices() {
+            return registrations.stream()
+                    .map(ActionServiceRegistration::actionService)
+                    .collect(toSet());
+        }
+    }
+
+    private record ActionServiceRegistration(
+            Class<? extends ThingActions> actionService,
+            ThingActions service,
+            ServiceRegistration<ThingActions> registration) {
+        private void unregister() {
+            service.dispose();
+            registration.unregister();
+        }
     }
 }
